add view custom section
