+++ custom_sections/serializer.py
@@ -1,0 +1,27 @@
+from rest_framework import serializers
+from product.serializer import ProductSerializer
+from .models import CustomSection
+
+class CustomSectionSerializer(serializers.ModelSerializer):
+    products = ProductSerializer(many=True, read_only=True, source='onlyGetSixProducts')
+    
+    class Meta:
+        model = CustomSection
+        fields = ['id', 'title', 'slug', 'products']
+        extra_kwargs = {'products': {'required': False, 'read_only': True}}
+        lookup_field = 'slug'
+        extra_kwargs = {
+            'url': {'lookup_field': 'slug'}
+        }
+
+class CustomSectionDetailSerializer(serializers.ModelSerializer):
+    products = ProductSerializer(many=True, read_only=True)
+    
+    class Meta:
+        model = CustomSection
+        fields = ['id', 'title', 'slug', 'products']
+        extra_kwargs = {'products': {'required': False, 'read_only': True}}
+        lookup_field = 'slug'
+        extra_kwargs = {
+            'url': {'lookup_field': 'slug'}
+        }
